refactor(lang): standardize adapter file names
