+++ internal/lang/php/adapter.go
@@ -1,14 +1,9 @@
 package php
 
 import (
-	"context"
-	"errors"
-	"io/fs"
-	"path/filepath"
 	"regexp"
 	"strings"
 
-	"github.com/ben-ranford/lopper/internal/lang/shared"
 	"github.com/ben-ranford/lopper/internal/language"
 )
 
@@ -16,79 +11,10 @@ type Adapter struct {
 	language.AdapterLifecycle
 }
 
-const (
-	composerJSONName = "composer.json"
-	composerLockName = "composer.lock"
-	maxDetectFiles   = 1024
-	maxScanFiles     = 2048
-)
-
 func NewAdapter() *Adapter {
 	adapter := &Adapter{}
 	adapter.AdapterLifecycle = language.NewAdapterLifecycle("php", []string{"php8", "php7"}, adapter.DetectWithConfidence)
 	return adapter
-}
-
-func (a *Adapter) DetectWithConfidence(ctx context.Context, repoPath string) (language.Detection, error) {
-	repoPath = shared.DefaultRepoPath(repoPath)
-	detection := language.Detection{}
-	roots := make(map[string]struct{})
-
-	if err := applyPHPRootSignals(repoPath, &detection, roots); err != nil {
-		return language.Detection{}, err
-	}
-
-	visited := 0
-	err := filepath.WalkDir(repoPath, func(path string, entry fs.DirEntry, walkErr error) error {
-		if walkErr != nil {
-			return walkErr
-		}
-		if ctx != nil && ctx.Err() != nil {
-			return ctx.Err()
-		}
-		return walkPHPDetectionEntry(path, entry, roots, &detection, &visited, maxDetectFiles)
-	})
-	if err != nil && !errors.Is(err, fs.SkipAll) {
-		return language.Detection{}, err
-	}
-
-	return shared.FinalizeDetection(repoPath, detection, roots), nil
-}
-
-func applyPHPRootSignals(repoPath string, detection *language.Detection, roots map[string]struct{}) error {
-	return shared.ApplyRootSignals(repoPath, phpRootSignals, detection, roots)
-}
-
-var phpRootSignals = []shared.RootSignal{
-	{Name: composerJSONName, Confidence: 60},
-	{Name: composerLockName, Confidence: 30},
-}
-
-func walkPHPDetectionEntry(path string, entry fs.DirEntry, roots map[string]struct{}, detection *language.Detection, visited *int, maxFiles int) error {
-	if entry.IsDir() {
-		if shouldSkipDir(entry.Name()) {
-			return filepath.SkipDir
-		}
-		return nil
-	}
-
-	*visited++
-	if *visited > maxFiles {
-		return fs.SkipAll
-	}
-
-	switch strings.ToLower(entry.Name()) {
-	case composerJSONName, composerLockName:
-		detection.Matched = true
-		detection.Confidence += 12
-		roots[filepath.Dir(path)] = struct{}{}
-	}
-
-	if strings.EqualFold(filepath.Ext(path), ".php") {
-		detection.Matched = true
-		detection.Confidence += 2
-	}
-	return nil
 }
 
 func normalizeDependencyID(value string) string {
