+++ internal/lang/rust/detection.go
