+++ internal/lang/swift/detection.go
